AMP-9467
git-svn-id: file:///svn/branches/AMP_1_15_RELEASE/amp@13545 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ WEB-INF/src/org/digijava/module/aim/action/SaveActivity.java
@@ -622,7 +622,7 @@ public class SaveActivity extends Action {
 						errors.add("sector", new ActionMessage("error.aim.addActivity.sectorEmpty", TranslatorWorker.translateText("Please add a sector",locale,siteId)));
 					}
 					else{
-						int primaryPrc=0, secondaryPrc=0, tertiaryPrc=0;
+						float primaryPrc=0, secondaryPrc=0, tertiaryPrc=0;
 						boolean hasPrimarySectorsAdded=false, hasSecondarySectorsAdded=false, hasTertiarySectorsAdded=false;
 						
 						Iterator<ActivitySector> secPerItr = eaForm.getSectors().getActivitySectors().iterator();
@@ -642,9 +642,9 @@ public class SaveActivity extends Action {
 							// sector percentage is not a number
 							else {
 								try {
-									if("Primary".equals(config.getName())) primaryPrc+=actSect.getSectorPercentage().intValue();
-									if("Secondary".equals(config.getName())) secondaryPrc+=actSect.getSectorPercentage().intValue();
-                                    if("Tertiary".equals(config.getName())) tertiaryPrc+=actSect.getSectorPercentage().intValue();
+									if("Primary".equals(config.getName())) primaryPrc+=actSect.getSectorPercentage().floatValue();
+									if("Secondary".equals(config.getName())) secondaryPrc+=actSect.getSectorPercentage().floatValue();
+                                    if("Tertiary".equals(config.getName())) tertiaryPrc+=actSect.getSectorPercentage().floatValue();
 								} catch (NumberFormatException nex) {
 									errors.add("sectorPercentageNonNumeric",
 											new ActionMessage("error.aim.addActivity.sectorPercentageNonNumeric", TranslatorWorker.translateText("Sector percentage must be numeric",locale,siteId)));
